john-oliver/update-ui-3 review and improvements. close PR 34
