--- app/backend/src/main/java/com/microsoft/openai/samples/rag/chat/controller/ChatResponse.java
@@ -1,6 +1,0 @@
-package com.microsoft.openai.samples.rag.chat.controller;
-
-import java.util.List;
-
-public record ChatResponse(List<ResponseChoice> choices) {
-}
