+++ app/backend/src/main/java/com/microsoft/openai/samples/rag/controller/ChatAppRequestContext.java
@@ -1,4 +1,4 @@
-package com.microsoft.openai.samples.rag.chat.controller;
+package com.microsoft.openai.samples.rag.controller;
 
 public record ChatAppRequestContext(ChatAppRequestOverrides overrides) {
 }
