+++ app/backend/src/main/java/com/microsoft/openai/samples/rag/controller/ResponseChoice.java
@@ -1,4 +1,4 @@
-package com.microsoft.openai.samples.rag.chat.controller;
+package com.microsoft.openai.samples.rag.controller;
 
 public record ResponseChoice(int index, ResponseMessage message, ResponseContext context) {
 }
