+++ app/backend/src/main/java/com/microsoft/openai/samples/rag/controller/ResponseContext.java
@@ -1,4 +1,4 @@
-package com.microsoft.openai.samples.rag.chat.controller;
+package com.microsoft.openai.samples.rag.controller;
 
 import java.util.List;
